Integration Compatibility + Tech Progress
The compatibility dialog has been visually added and scripted to show when in simple mode but hidden in tech mode. It is not yet functional on detections.

Separate tech platform dialogs have been created for a number of compatible platforms and progress on their setup but no function as of yet.