New MT Detection, Saving Dialog, Chip Linking Groundwork, Bug Fixes
The MT will now display in its own panel on the right side of the window. When more than one MT are detected, it will display a tooltip on rollover with the MIDs attached to said detections.

Created the future-use JS file for setting up Wiki links off of the ease of installation chips.

When typing in the clipboard, the clear button will now display that it is saving your data to naturally reflect use.

Removed Awin cookie detection as it's no longer possible to detect first-party cookies.

Fixed the display of some UI elements.

Updated font family in use.