Analytics Fixes, Chip Linking, Code Cleanup
Analytics was reading event interaction with the actual page rather than just the extension. Renamed content-scripts.js to popup.js and moved it to only load with in the head of the popup.

Chips now link to their relevant documentation; need to figure out a better way to link to all documents.

Removed excess code from un-adopted features. These may be added back in the future but will remain off of the build for now.